[change] move test file directory to 'src/tests/**.test.ts'
Because put all typescript files in src directory
